+++ Changelog.md
@@ -9,7 +9,11 @@
 - Allow users to customize video-compare settings in the Settings menu
 - Force kill trtexec and vspipe processes when beginning workflow processing
 - Add MC_Degrain filters
-- Change to vs-mlrt version 15.14 from 15.13 RTX. This should enable GTX GPUs to work
+- Change to vs-mlrt version 15.14 from 15.13 RTX
+- Add detection for vs-mlrt version changing (will not take effect in this release)
+- Add BF16 toggle when building TensorRT models
+- Add automatic static + shape detection when building TensorRT models
+- Add update system for vs-mlrt plugin
 
 ## 0.10.2
 - Implement segment selection. Users can now select a small segment of a video to process and preview!